Move inference server out of common
